feat(*): rename /shared to /_lit-styles
