+++ 2nd-gen/packages/swc/stylesheets/_lit-styles/linear-progress-base.css
