+++ 2nd-gen/packages/swc/stylesheets/_lit-styles/opacity-checkerboard.css
